fixedbugsforcourses

+++ src/courses/courses.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
-import { CourseDto } from './dto/main';
-import { Prisma, Courses } from '@prisma/client';
+import { CourseDto, UpdateCourseDto, DeletedCourseDto } from './dto/main';
+import { Courses } from '@prisma/client';
 
 @Injectable()
 export class CoursesService {
@@ -11,8 +11,11 @@ export class CoursesService {
     return this.prisma.courses.create({ data: courseData });
   }
 
-  async findOne(id: number): Promise<Courses | null> {
-    const course = await this.prisma.courses.findFirst({ where: { id: id } });
+  async findOne(id: number | string): Promise<Courses | null> {
+    const courseid = Number(id);
+    const course = await this.prisma.courses.findFirst({
+      where: { id: courseid },
+    });
 
     if (!course) {
       throw new NotFoundException('Course not found');
@@ -21,21 +24,22 @@ export class CoursesService {
     return course;
   }
 
-  async update(id: number, courseData: CourseDto): Promise<Courses | null> {
+  async update(
+    id: number,
+    courseData: UpdateCourseDto,
+  ): Promise<Courses | null> {
     return this.prisma.courses.update({ where: { id }, data: courseData });
   }
 
-  async remove(id: number): Promise<Courses | null> {
-    const deletedCourse = await this.prisma.courses.delete({
-      where: {
-        id: id,
-      },
-    });
-
-    if (!deletedCourse) {
-      throw new NotFoundException('Course not found');
+  async remove(id: number | string): Promise<DeletedCourseDto | null> {
+    const courseId = Number(id);
+    const delcourse = await this.findOne(courseId);
+    if (!delcourse) {
+      throw new NotFoundException('Not found');
     }
-
-    return deletedCourse;
+    await this.prisma.courses.delete({
+      where: { id: courseId },
+    });
+    return delcourse;
   }
 }
